Align manifest parsing with the upstream corpus
The vendored upstream corpus (next commit) exposed compatibility gaps in
strict parsing, each now covered: property names match case-insensitively
with insignificant surrounding whitespace the way the upstream JSON layer
does; precedence accepts string and number forms; rename accepts the map
form in addition to the array form; generatorVersions accepts a single
version string; post action args accept string arrays, joined with ';';
condition properties accept boolean literals; and the legacy onlyIf,
generated-symbol defaultValue, and post action id /
applyFileRenamesToArgs / applyFileRenamesToManualInstructions properties
are accepted. The anchored-replacement semantics of onlyIf remain
unimplemented and are recorded as a deviation in SPECIFICATION.md.

+++ Source/Templating/TemplateConfig.cs
@@ -52,7 +52,7 @@ public record TemplateConfig
     public string? GroupIdentity { get; init; }
 
     /// <summary>Gets the precedence within the group identity.</summary>
-    public int Precedence { get; init; }
+    public double Precedence { get; init; }
 
     /// <summary>Gets the default name for instantiations.</summary>
     public string? DefaultName { get; init; }
